+++ packages/components-react/src/pages/Action.tsx
@@ -126,6 +126,7 @@ export function Action() {
           <div className="playground light spacing-inline">
             <ButtonPure weight={'thin'}>Thin</ButtonPure>
             <ButtonPure weight={'regular'}>Regular</ButtonPure>
+            <ButtonPure weight={'semibold'}>Semibold</ButtonPure>
             <ButtonPure weight={'bold'}>Bold</ButtonPure>
           </div>
           <div className="playground light spacing-inline">
@@ -133,6 +134,16 @@ export function Action() {
           </div>
           <div className="playground light spacing-inline">
             <ButtonPure style={{width: '240px'}}>Lorem ipsum dolor sit amet, consetetur sadipscing</ButtonPure>
+          </div>
+          <div className="playground light spacing-inline">
+            <ButtonPure weight={'semibold'} size={'medium'}>
+              Some label
+              <p slot="subline">Some subline</p>
+            </ButtonPure>
+            <ButtonPure weight={'semibold'} size={'medium'} disabled={true}>
+              Some label
+              <p slot="subline">Some subline</p>
+            </ButtonPure>
           </div>
           <Divider/>
         </GridItem>
@@ -154,6 +165,7 @@ export function Action() {
           <div className="playground dark spacing-inline">
             <ButtonPure weight={'thin'} theme={'dark'}>Thin</ButtonPure>
             <ButtonPure weight={'regular'} theme={'dark'}>Regular</ButtonPure>
+            <ButtonPure weight={'semibold'} theme={'dark'}>Semibold</ButtonPure>
             <ButtonPure weight={'bold'} theme={'dark'}>Bold</ButtonPure>
           </div>
           <div className="playground dark spacing-inline">
@@ -162,6 +174,16 @@ export function Action() {
           <div className="playground dark spacing-inline">
             <ButtonPure theme={'dark'} style={{width: '240px'}}>Lorem ipsum dolor sit amet, consetetur
               sadipscing</ButtonPure>
+          </div>
+          <div className="playground dark spacing-inline">
+            <ButtonPure theme={'dark'} weight={'semibold'} size={'medium'}>
+              Some label
+              <p slot="subline">Some subline</p>
+            </ButtonPure>
+            <ButtonPure theme={'dark'} weight={'semibold'} size={'medium'} disabled={true}>
+              Some label
+              <p slot="subline">Some subline</p>
+            </ButtonPure>
           </div>
           <Divider/>
         </GridItem>